refactor(exercise): move exercise menu item unwrap to selection component
